requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
dj-static==0.0.6
Django==1.9.7
django-ckeditor==5.0.3
django-cors-headers==1.1.0
django-cors-middleware==1.2.0
django-extensions==1.6.6
django-rest-auth==0.7.0
djangorestframework==3.3.3
gnureadline==6.3.3
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
static3==0.6.1
traitlets==4.2.1
Werkzeug==0.11.9
